refactored grid component/s
