+++ app/cdap/components/GridTable/GridHeaderCell/types.ts
@@ -1,0 +1,4 @@
+export interface IGridHeaderCellProps {
+  label: string;
+  types: string[];
+}
