+++ app/cdap/components/GridTable/GridKPICell/types.ts
@@ -1,0 +1,6 @@
+export interface IGridKPICellProps {
+  metricData: {
+    name: string;
+    values: Array<{ label: string; count: number }>;
+  };
+}
