+++ app/cdap/components/GridTable/GridTextCell/types.ts
@@ -1,0 +1,3 @@
+export interface IGridTextCellProps {
+  cellValue: string;
+}
