+++ app/cdap/components/GridTable/Typography/types.ts
@@ -1,0 +1,4 @@
+export interface ITypographyTextComponentProps {
+  className: string;
+  label: string;
+}
